Creating Health UI in Alex Gameplay Scene.
added code in Stats (CoreComponents) on lines 3, 16, 17, and 36. Needed UnityEngine.UI header, then lines 16, 17 are for the use of images and gradient. Line 36 allows use of a health gradient (color goes from yellow to red as health lowers) , but doesn't seem to be appearing on the Unity side for some reason. A red fill behind the green health bar is in place for now.